implementing import

+++ src/DotClassReader/Interface.cpp
@@ -19,7 +19,7 @@ std::vector<int> Interface::getITF() { return itf; }
 int Interface::itfCount() { return interface_count; }
 
 void Interface::seek() {
-    interface_count = getInfo(file, 2);
+    interface_count = (int)getInfo(file, 2);
     for (int i = 0; i < interface_count; i++) {
         itf.push_back(getInfo(file, 2));
     }
